+++ CHANGELOG.md
@@ -3,7 +3,13 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Incomplete type conversion of int, decimal, time, date and datetime.
+
 [4.22.0] - 2017-02-17
+---------------------
 ##### Added
 - Support for lang attributes on choice labels in secondary instances.